arrange directory structure
